+++ src/account/operations/syncing/foundries.rs
@@ -5,7 +5,7 @@ use std::collections::HashSet;
 
 use iota_client::block::output::{FoundryId, Output};
 
-use crate::account::handle::AccountHandle;
+use crate::{account::handle::AccountHandle, task};
 
 impl AccountHandle {
     pub(crate) async fn request_and_store_foundry_outputs(&self, foundry_ids: HashSet<FoundryId>) -> crate::Result<()> {
@@ -22,7 +22,7 @@ impl AccountHandle {
 
             let client = self.client.clone();
             tasks.push(async move {
-                tokio::spawn(async move {
+                task::spawn(async move {
                     match client.foundry_output_id(foundry_id).await {
                         Ok(output_id) => Ok(Some(client.get_output(&output_id).await?)),
                         Err(iota_client::Error::NotFound(_)) => Ok(None),