+++ src/account_manager/builder.rs
@@ -19,11 +19,8 @@ use crate::events::EventEmitter;
 #[cfg(all(feature = "storage", not(feature = "rocksdb")))]
 use crate::storage::adapter::memory::Memory;
 #[cfg(feature = "storage")]
-use crate::{
-    account::handle::AccountHandle,
-    storage::{constants::default_storage_path, manager::ManagerStorage},
-};
-use crate::{account_manager::AccountManager, ClientOptions, Error};
+use crate::storage::{constants::default_storage_path, manager::ManagerStorage};
+use crate::{account::handle::AccountHandle, account_manager::AccountManager, ClientOptions};
 
 #[derive(Debug, Clone, Serialize, Deserialize, Default)]
 /// Builder for the account manager.
@@ -138,7 +135,7 @@ impl AccountManagerBuilder {
         #[cfg(feature = "storage")]
         let read_manager_builder = match storage_manager.lock().await.get_account_manager_data().await {
             Ok(x) => Some(x),
-            Err(Error::RecordNotFound(_)) => None,
+            Err(crate::Error::RecordNotFound(_)) => None,
             Err(e) => return Err(e),
         };
         #[cfg(not(feature = "storage"))]